# Optimize Ajv benchmark